# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Indonesia Pipeline Dashboard",
    "description": "Serves a live dashboard of Ninja Van's Indonesia sales opportunity pipeline (composition, aging, stage bottlenecks, 3-month forecast, and team rollups), fetched from the Ninja Van SalesCRM API and cached for up to 5 minutes.",
    "version": "3.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Readiness probe used by the platform to confirm the backend is up.",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "The backend is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline": {
      "get": {
        "summary": "Returns aggregated Indonesia opportunity pipeline metrics fetched live from the CRM (cached up to 5 minutes).",
        "operationId": "get_pipeline",
        "responses": {
          "200": {
            "description": "Aggregated pipeline metrics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "open_pipeline": { "type": "integer" },
                    "open_active": { "type": "integer" },
                    "open_future": { "type": "integer" },
                    "closed_won": { "type": "integer" },
                    "closed_lost": { "type": "integer" },
                    "closed_total": { "type": "integer" },
                    "win_rate": { "type": "number" },
                    "revenue_potential_mth": { "type": "number" },
                    "revenue_committed_mth": { "type": "number" },
                    "open_stages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "closed_stages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "product_lines": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "service_levels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "service_blank": { "type": "integer" },
                    "owners": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "aging_buckets": {
                      "type": "array",
                      "description": "Open deals bucketed by days since creation.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "stage_duration": {
                      "type": "array",
                      "description": "Average days open deals have sat in each current stage, longest first.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "avg_days": { "type": "number" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "stage_duration_buckets": {
                      "type": "array",
                      "description": "Open deals bucketed by days since their last stage change.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "stage_duration_grid": {
                      "type": "array",
                      "description": "For each stage, its open deals broken down by days-since-last-stage-change bucket.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "stage": { "type": "string" },
                          "total": { "type": "integer" },
                          "buckets": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "label": { "type": "string" },
                                "count": { "type": "integer" }
                              }
                            }
                          }
                        }
                      }
                    },
                    "aging_grid": {
                      "type": "array",
                      "description": "For each days-since-creation bucket, open deals broken down by current stage.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string" },
                          "total": { "type": "integer" },
                          "stages": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": { "type": "string" },
                                "count": { "type": "integer" }
                              }
                            }
                          }
                        }
                      }
                    },
                    "forecast": {
                      "type": "array",
                      "description": "Open deals by expected close date: a leading 'overdue' bucket (close date already passed), then one bucket per month from the current month through the furthest expected close date in the data. Each bucket is broken down by stage and revenue.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": { "type": "string", "example": "2026-09", "description": "\"overdue\" or a YYYY-MM month key." },
                          "count": { "type": "integer" },
                          "revenue": { "type": "number" },
                          "stages": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": { "type": "string" },
                                "count": { "type": "integer" },
                                "revenue": { "type": "number" }
                              }
                            }
                          }
                        }
                      }
                    },
                    "forecast_undated_count": {
                      "type": "integer",
                      "description": "Open deals with no expected close date, excluded from the forecast."
                    },
                    "created_by_month": {
                      "type": "object",
                      "description": "Opportunities created per month, rolled up by Manager and Sales Head from the Active-rep team roster.",
                      "properties": {
                        "months": { "type": "array", "items": { "type": "string" } },
                        "by_manager": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": { "type": "string" },
                              "counts": { "type": "object", "additionalProperties": { "type": "integer" } },
                              "total": { "type": "integer" }
                            }
                          }
                        },
                        "by_sales_head": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": { "type": "string" },
                              "counts": { "type": "object", "additionalProperties": { "type": "integer" } },
                              "total": { "type": "integer" }
                            }
                          }
                        },
                        "unmapped_owner_count": { "type": "integer" }
                      }
                    },
                    "lookback_days": { "type": "integer" },
                    "snapshot_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "502": { "description": "The CRM rejected the API key (expired or revoked)." },
          "503": { "description": "CRM_API_KEY is not configured on this app." }
        }
      }
    },
    "/{full_path}": {
      "get": {
        "summary": "Serves the Indonesia opportunity pipeline dashboard page for any path.",
        "operationId": "serve_dashboard",
        "parameters": [
          {
            "name": "full_path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The dashboard HTML page.",
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}